Move new step logic into optimize/step.py
